move generateUUID to common utils (still it's better to switch to npm package)
